+++ TinkoffASDKUI/TinkoffASDKUI/UIComponents/EmailView/Assembly/IEmailViewPresenterAssembly.swift
@@ -1,0 +1,10 @@
+//
+//  IEmailViewPresenterAssembly.swift
+//  TinkoffASDKUI
+//
+//  Created by Aleksandr Pravosudov on 23.05.2023.
+//
+
+protocol IEmailViewPresenterAssembly {
+    func build(customerEmail: String, output: IEmailViewPresenterOutput) -> IEmailViewOutput
+}